+++ sdk/imagebuilder/azure-resourcemanager-imagebuilder/src/main/java/com/azure/resourcemanager/imagebuilder/implementation/TriggerImpl.java
@@ -60,20 +60,14 @@ public final class TriggerImpl implements Trigger, Trigger.Definition, Trigger.U
     }
 
     public Trigger create() {
-        this.innerObject =
-            serviceManager
-                .serviceClient()
-                .getTriggers()
-                .createOrUpdate(resourceGroupName, imageTemplateName, triggerName, this.innerModel(), Context.NONE);
+        this.innerObject = serviceManager.serviceClient().getTriggers().createOrUpdate(resourceGroupName,
+            imageTemplateName, triggerName, this.innerModel(), Context.NONE);
         return this;
     }
 
     public Trigger create(Context context) {
-        this.innerObject =
-            serviceManager
-                .serviceClient()
-                .getTriggers()
-                .createOrUpdate(resourceGroupName, imageTemplateName, triggerName, this.innerModel(), context);
+        this.innerObject = serviceManager.serviceClient().getTriggers().createOrUpdate(resourceGroupName,
+            imageTemplateName, triggerName, this.innerModel(), context);
         return this;
     }
 
@@ -88,20 +82,14 @@ public final class TriggerImpl implements Trigger, Trigger.Definition, Trigger.U
     }
 
     public Trigger apply() {
-        this.innerObject =
-            serviceManager
-                .serviceClient()
-                .getTriggers()
-                .createOrUpdate(resourceGroupName, imageTemplateName, triggerName, this.innerModel(), Context.NONE);
+        this.innerObject = serviceManager.serviceClient().getTriggers().createOrUpdate(resourceGroupName,
+            imageTemplateName, triggerName, this.innerModel(), Context.NONE);
         return this;
     }
 
     public Trigger apply(Context context) {
-        this.innerObject =
-            serviceManager
-                .serviceClient()
-                .getTriggers()
-                .createOrUpdate(resourceGroupName, imageTemplateName, triggerName, this.innerModel(), context);
+        this.innerObject = serviceManager.serviceClient().getTriggers().createOrUpdate(resourceGroupName,
+            imageTemplateName, triggerName, this.innerModel(), context);
         return this;
     }
 
@@ -114,22 +102,14 @@ public final class TriggerImpl implements Trigger, Trigger.Definition, Trigger.U
     }
 
     public Trigger refresh() {
-        this.innerObject =
-            serviceManager
-                .serviceClient()
-                .getTriggers()
-                .getWithResponse(resourceGroupName, imageTemplateName, triggerName, Context.NONE)
-                .getValue();
+        this.innerObject = serviceManager.serviceClient().getTriggers()
+            .getWithResponse(resourceGroupName, imageTemplateName, triggerName, Context.NONE).getValue();
         return this;
     }
 
     public Trigger refresh(Context context) {
-        this.innerObject =
-            serviceManager
-                .serviceClient()
-                .getTriggers()
-                .getWithResponse(resourceGroupName, imageTemplateName, triggerName, context)
-                .getValue();
+        this.innerObject = serviceManager.serviceClient().getTriggers()
+            .getWithResponse(resourceGroupName, imageTemplateName, triggerName, context).getValue();
         return this;
     }