+++ sdk/imagebuilder/azure-resourcemanager-imagebuilder/CHANGELOG.md
@@ -1,14 +1,33 @@
 # Release History
 
-## 1.0.0-beta.5 (Unreleased)
+## 1.0.0 (2023-12-19)
+
+- Azure Resource Manager ImageBuilder client library for Java. This package contains Microsoft Azure SDK for ImageBuilder Management SDK. Azure Virtual Machine Image Builder Client. Package tag package-2023-07. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Features Added
 
-### Breaking Changes
+* `models.ImageTemplateUpdateParametersProperties` was added
 
-### Bugs Fixed
+* `models.OnBuildError` was added
 
-### Other Changes
+* `models.ImageTemplatePropertiesErrorHandling` was added
+
+#### `models.ImageTemplateUpdateParameters` was modified
+
+* `withProperties(models.ImageTemplateUpdateParametersProperties)` was added
+* `properties()` was added
+
+#### `models.ImageTemplate$Update` was modified
+
+* `withProperties(models.ImageTemplateUpdateParametersProperties)` was added
+
+#### `models.ImageTemplate` was modified
+
+* `errorHandling()` was added
+
+#### `models.ImageTemplate$Definition` was modified
+
+* `withErrorHandling(models.ImageTemplatePropertiesErrorHandling)` was added
 
 ## 1.0.0-beta.4 (2023-03-21)